Added bird and game states

+++ FlappyBird/FlappyBird/PipeManager.cpp
@@ -81,7 +81,7 @@ std::vector<Pipe*> PipeManager::GetPipes() const
 
 void PipeManager::CreatePipe()
 {
-	GLfloat spawnY = rand() % (minSpawnY - maxSpawnY) + maxSpawnY;
+	GLfloat spawnY = (GLfloat)(rand() % (minSpawnY - maxSpawnY) + maxSpawnY);
 	Pipe* pipe = new Pipe(glm::vec2(Engine::SCREEN_WIDTH + 90.0f, spawnY), glm::vec2(90.0f, 500.0f), ResourceManager::GetTexture("pipe"), 200.0f, 80.0f);
 	
 	pipes.push_back(pipe);
